Test for existence of casc_agent_1_new before trying to delete it

+++ dockerrun.sh
@@ -164,7 +164,9 @@ if [ ! -d "$SCRIPT_DIR/paper_results/raw_data/casc_agent_1" ]; then
     mkdir -p "$SCRIPT_DIR/paper_results/raw_data"
     RESULTSBIND=""
     DUMMYCONTAINER=`docker run --rm -d --mount type=volume,dst=/home/carla/EnergyShield/models,volume-driver=local,volume-opt=type=none,volume-opt=o=bind,volume-opt=device="${SCRIPT_DIR}/paper_results/raw_data" energyshield:${localuser} $user -d $SERVER $CORES $PORTNUM $CARLAPORT $MPIHOSTS ""`
-    rm -rf "$SCRIPT_DIR/paper_results/raw_data/casc_agent_1_new"
+    if [ -d "$SCRIPT_DIR/paper_results/raw_data/casc_agent_1_new" ]; then
+        rm -rf "$SCRIPT_DIR/paper_results/raw_data/casc_agent_1_new"
+    fi
     DCONTAINERS=`docker container ls -a | grep energyshield:$localuser | sed -e "s/[ ].*//"`
     for CONT in $DCONTAINERS; do
       if [ ${DUMMYCONTAINER:0:12} = $CONT ]; then
